separate tick from ticker and fix holiday issue
